refactor: share blueprint helpers
